TILEMAPPING
Using Eris' environmental sprites (thank you!), I created a tilemap of most of what we will be working on
The tilemap has colliders, but it might need some work to make it work with player movement.
Got the Zora (water enemy) work with going underwater/resurfacing at a different location - still got to work on the animation part, and that enemies fireballs
Camera movement needs to be reworked to work with the stat bar at the top of the screen.